Add some textboxes, and other features to get ready to add user controls to the dataset.
