+++ Inventory/InventoryDataSet.xsc
@@ -5,5 +5,20 @@
      the code is regenerated.
 </autogenerated>-->
 <DataSetUISetting Version="1.00" xmlns="urn:schemas-microsoft-com:xml-msdatasource">
-  <TableUISettings />
+  <TableUISettings>
+    <TableUISetting Name="Inventory">
+      <ColumnUISettings>
+        <ColumnUISetting Name="ID">
+          <ControlSettings><ControlSetting ArtifactName="Microsoft:System.Windows.Forms:Form" xmlns="urn:schemas-microsoft-com:xml-msdatasource">
+              <BindableControlInfo Name="TextBox" Type="System.Windows.Forms.TextBox" AssemblyName="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+            </ControlSetting></ControlSettings>
+        </ColumnUISetting>
+        <ColumnUISetting Name="Item">
+          <ControlSettings><ControlSetting ArtifactName="Microsoft:System.Windows.Forms:Form" xmlns="urn:schemas-microsoft-com:xml-msdatasource">
+              <BindableControlInfo Name="TextBox" Type="System.Windows.Forms.TextBox" AssemblyName="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+            </ControlSetting></ControlSettings>
+        </ColumnUISetting>
+      </ColumnUISettings>
+    </TableUISetting>
+  </TableUISettings>
 </DataSetUISetting>
