+++ Inventory/InventoryDataSet.xsd
@@ -125,6 +125,23 @@ WHERE        (Employee = ?)</CommandText>
                   </DbCommand>
                 </SelectCommand>
               </DbSource>
+              <DbSource ConnectionRef="InventoryConnectionString (Settings)" DbObjectName="" DbObjectType="Unknown" GenerateShortCommands="true" GeneratorSourceName="InsertQuery" Modifier="Public" Name="InsertQuery" QueryType="NoData" ScalarCallRetval="System.Object, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" UseOptimisticConcurrency="true" UserGetMethodName="GetDataBy2" UserSourceName="InsertQuery">
+                <InsertCommand>
+                  <DbCommand CommandType="Text" ModifiedByUser="true">
+                    <CommandText>INSERT INTO `Inventory` (`Item`, `Employee`, `Department`, `Serial#`, `ItemTag#`, `Installed`, `ReplaceBy`, `Active`) VALUES (?, ?, ?, ?, ?, ?, ?, ?)</CommandText>
+                    <Parameters>
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param1" ColumnName="Item" DataSourceName="Inventory" DataTypeServer="DbType.StringFixedLength(255)" DbType="String" Direction="Input" ParameterName="Item" Precision="0" ProviderType="WChar" Scale="0" Size="255" SourceColumn="Item" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param2" ColumnName="Employee" DataSourceName="Inventory" DataTypeServer="DbType.StringFixedLength(255)" DbType="String" Direction="Input" ParameterName="Employee" Precision="0" ProviderType="WChar" Scale="0" Size="255" SourceColumn="Employee" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param3" ColumnName="Department" DataSourceName="Inventory" DataTypeServer="DbType.StringFixedLength(255)" DbType="String" Direction="Input" ParameterName="Department" Precision="0" ProviderType="WChar" Scale="0" Size="255" SourceColumn="Department" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param4" ColumnName="Serial#" DataSourceName="Inventory" DataTypeServer="DbType.StringFixedLength(255)" DbType="String" Direction="Input" ParameterName="Serial_" Precision="0" ProviderType="WChar" Scale="0" Size="255" SourceColumn="Serial#" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param5" ColumnName="ItemTag#" DataSourceName="Inventory" DataTypeServer="DbType.StringFixedLength(255)" DbType="String" Direction="Input" ParameterName="ItemTag_" Precision="0" ProviderType="WChar" Scale="0" Size="255" SourceColumn="ItemTag#" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param6" ColumnName="Installed" DataSourceName="Inventory" DataTypeServer="DbType.Date" DbType="Date" Direction="Input" ParameterName="Installed" Precision="0" ProviderType="Date" Scale="0" Size="0" SourceColumn="Installed" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="true" AutogeneratedName="Param7" ColumnName="ReplaceBy" DataSourceName="Inventory" DataTypeServer="DbType.Date" DbType="Date" Direction="Input" ParameterName="ReplaceBy" Precision="0" ProviderType="Date" Scale="0" Size="0" SourceColumn="ReplaceBy" SourceColumnNullMapping="false" SourceVersion="Current" />
+                      <Parameter AllowDbNull="false" AutogeneratedName="Param8" ColumnName="Active" DataSourceName="Inventory" DataTypeServer="DbType.Boolean(2)" DbType="Boolean" Direction="Input" ParameterName="Active" Precision="0" ProviderType="Boolean" Scale="0" Size="2" SourceColumn="Active" SourceColumnNullMapping="false" SourceVersion="Current" />
+                    </Parameters>
+                  </DbCommand>
+                </InsertCommand>
+              </DbSource>
             </Sources>
           </TableAdapter>
         </Tables>
@@ -135,7 +152,7 @@ WHERE        (Employee = ?)</CommandText>
   <xs:element name="InventoryDataSet" msdata:IsDataSet="true" msdata:UseCurrentLocale="true" msprop:EnableTableAdapterManager="True" msprop:Generator_DataSetName="InventoryDataSet" msprop:Generator_UserDSName="InventoryDataSet">
     <xs:complexType>
       <xs:choice minOccurs="0" maxOccurs="unbounded">
-        <xs:element name="Inventory" msprop:Generator_TableClassName="InventoryDataTable" msprop:Generator_TableVarName="tableInventory" msprop:Generator_RowChangedName="InventoryRowChanged" msprop:Generator_TablePropName="Inventory" msprop:Generator_RowDeletingName="InventoryRowDeleting" msprop:Generator_RowChangingName="InventoryRowChanging" msprop:Generator_RowEvHandlerName="InventoryRowChangeEventHandler" msprop:Generator_RowDeletedName="InventoryRowDeleted" msprop:Generator_RowClassName="InventoryRow" msprop:Generator_UserTableName="Inventory" msprop:Generator_RowEvArgName="InventoryRowChangeEvent">
+        <xs:element name="Inventory" msprop:Generator_TableClassName="InventoryDataTable" msprop:Generator_TableVarName="tableInventory" msprop:Generator_TablePropName="Inventory" msprop:Generator_RowDeletingName="InventoryRowDeleting" msprop:Generator_RowChangingName="InventoryRowChanging" msprop:Generator_RowEvHandlerName="InventoryRowChangeEventHandler" msprop:Generator_RowDeletedName="InventoryRowDeleted" msprop:Generator_UserTableName="Inventory" msprop:Generator_RowChangedName="InventoryRowChanged" msprop:Generator_RowEvArgName="InventoryRowChangeEvent" msprop:Generator_RowClassName="InventoryRow">
           <xs:complexType>
             <xs:sequence>
               <xs:element name="ID" msdata:AutoIncrement="true" msdata:AutoIncrementSeed="-1" msdata:AutoIncrementStep="-1" msprop:Generator_ColumnVarNameInTable="columnID" msprop:Generator_ColumnPropNameInRow="ID" msprop:Generator_ColumnPropNameInTable="IDColumn" msprop:Generator_UserColumnName="ID" type="xs:int" />
